+++ napari/plugins/tests/fixtures/napari_test_plugin.py
@@ -6,8 +6,8 @@ import pluggy
 napari_hook_implementation = pluggy.HookimplMarker("napari")
 
 
-def reader_function():
-    pass
+def reader_function(path):
+    return True
 
 
 @napari_hook_implementation
